backend/requirements.txt
fastapi
uvicorn
pydantic
websockets
httpx
supabase
python-dotenv
